acceptance_spec_helper.rb: Move the file to spec/spec_helpers directory
In order to be able to execute RSpec tests individually we
have moved this file to spec/spec_helpers directory.

This is needed because when executing specific RSpec tests,
or specific tests in a inner subdirectory in the spec directory,
the spec_helper.rb file is not automatically included by Rake as it
is done when the spec directory is executed entirely by RSpec. We
have moved the file to spec/spec_helpers because the spec/spec_helpers
files are included by the spec/spec_helper.rb file through a
require_relative directive, and from Rake we set the RSpec tasks
options to require the spec_helper.rb file.

All the acceptance/api/* files that referenced
the acceptance_spec_helper.rb have been updated accordingly
