+++ spec/acceptance/api/internal/alert_limits_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../../acceptance_spec_helper'
+require_relative '../../../spec_helpers/acceptance_spec_helper'
 
 resource 'Alert limits' do
   header 'Accept', 'application/json'
