+++ spec/acceptance/api/internal/application_keys_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../../acceptance_spec_helper'
+require_relative '../../../spec_helpers/acceptance_spec_helper'
 
 resource 'Application keys' do
   header 'Accept', 'application/json'
