+++ spec/acceptance/api/internal/applications_api_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../../acceptance_spec_helper'
+require_relative '../../../spec_helpers/acceptance_spec_helper'
 
 resource 'Applications (prefix: /services/:service_id/applications)' do
   header 'Accept', 'application/json'
