+++ spec/acceptance/api/internal/errors_api_spec.rb
@@ -1,5 +1,5 @@
 require 'timecop'
-require_relative '../../acceptance_spec_helper'
+require_relative '../../../spec_helpers/acceptance_spec_helper'
 
 resource 'Errors (prefix: /services/:service_id/errors)' do
   header 'Accept', 'application/json'
