+++ spec/acceptance/api/internal/events_api_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../../acceptance_spec_helper'
+require_relative '../../../spec_helpers/acceptance_spec_helper'
 
 resource 'Events' do
   header 'Accept', 'application/json'
