+++ spec/acceptance/api/internal/internal_api_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../../acceptance_spec_helper'
+require_relative '../../../spec_helpers/acceptance_spec_helper'
 
 resource 'Internal API (prefix: /internal)' do
   header 'Accept', 'application/json'
