+++ spec/acceptance/api/internal/service_tokens_api_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../../acceptance_spec_helper'
+require_relative '../../../spec_helpers/acceptance_spec_helper'
 
 resource 'Service Tokens (prefix: /service_tokens)' do
   header 'Accept', 'application/json'
