+++ spec/acceptance/api/internal/services_api_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../../acceptance_spec_helper'
+require_relative '../../../spec_helpers/acceptance_spec_helper'
 
 resource 'Services (prefix: /services)' do
   header 'Accept', 'application/json'
