+++ spec/acceptance/api/internal/transactions_api_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../../acceptance_spec_helper'
+require_relative '../../../spec_helpers/acceptance_spec_helper'
 
 resource 'Transactions (prefix: /services/:service_id/transactions)' do
   header 'Accept', 'application/json'
