+++ spec/acceptance/api/internal/usagelimits_api_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../../acceptance_spec_helper'
+require_relative '../../../spec_helpers/acceptance_spec_helper'
 
 resource 'UsageLimits (prefix: /services/:service_id/plans/:plan_id/usagelimits)' do
   header 'Accept', 'application/json'
