+++ spec/acceptance/api/internal/utilization_api_spec.rb
@@ -1,4 +1,4 @@
-require_relative '../../acceptance_spec_helper'
+require_relative '../../../spec_helpers/acceptance_spec_helper'
 
 resource 'Utilization (prefix: /services/:service_id/applications/:app_id/utilization)' do
   header 'Accept', 'application/json'
